Add a TweetFormatter
The TweetFormatter takes a string as input and ouputs and array of
strings of less than 280 characters.

Furthermore it numbers tweets if there is more than one so they appear
connected in a thread of tweets.

At the core of splitting a long string into tweetable substrings of no
more than 280 characters are the methods `String#rindex`[1] and
`String#slice!`[2].

`#rindex` takes a character and a max length as arguments. The character
is an empty space and the length the number of characters that fit in a
tweet. That is we find the last occurrence of an empty space within 280
characters of the string.

`#slice!` takes two arguments as well. A 0 to get the first element of
the string and then the last ocurrence of an empty space obtained by
`#rindex`. The caller (`input`) is mutated as the string is sliced by
our specifications.

By using these methods we can split a long input of text into smaller
chunks that work well as tweets. We split these on an empty space rather
than having to break up words in the middle.

In the end we get an array of strings where each string maps to one
tweet.

Thanks to this StackOverflow answer for suggesting to use `rindex`[3]

For the spec it was rather cumbersome to work with long strings that
span many columns. The squiggly heredoc is useful in this case, but still
introduced new lines which obstructed the string matching in tests.
Hence the solution of creating an array and joining it on whitespace
seemed appropriate[4].

[1] https://ruby-doc.org/core-2.6.1/String.html#method-i-rindex
[2] https://ruby-doc.org/core-2.6.1/String.html#method-i-slice-21
[3] https://stackoverflow.com/a/16114526/4928634
[4] https://stackoverflow.com/a/45757084/4928634