Add TwitterClient
The TwiterClient works with the Ruby Twitter gem and exposes one key
public method, `TwitterClient#send_tweets`.

The method iterates over an array of strings and sends a status update,
e.i. a tweet for each.

If the array contains just one string it should just send one tweet. In
that case the local variable `tweet` returns `nil`, so it's not in
response to any previous tweets.

If the array contains multiple strings the tweets should be connected in
a thread, which can be achieved by replying to the previous tweet. Hence
at the second time of iterating over the loop the variable `tweet`
should contain the previous `Twitter:Tweet` object[1]. By passing that
to the `#update` method the tweet is in reply to the previous tweet,
thereby creating a thread.

[1] https://www.rubydoc.info/gems/twitter/Twitter/Tweet