Create Tweeter to orchestrate the tweeting of a string
This adds a Tweeter class to orchestrate the basic feature of taking
input from STDIN and breaking the received string up into a
single or multiple individual strings that are tweeted and connected by
a thread.

The use of Ruby's `ARGF.read` allows the script to take input from a file
passed as a command line argument directly or by piping input into the
script.