Create main script to execute froom command line
This script calls the tweeter class. It is simple and untested.

Initializing the class and calling the `Twitter#tweet` instance method
directly outisde the class from within the same file caused Twitter to
be called each time Rspec was run, despite stubbing the API.

Some ways to call the script are:

* ./main.rb text.md
* cat text.md | ./main.rb
* echo "hello twitter" | ./main.rb

Note that `echo` will add a newline to the last tweet, if not used in combination with the
`-n` flag, e.g. `$ echo -n "hello world"`.