Add feature test
This tests the passing of a string to STDIN and the subsequent chopping
and tweeting of each sub-string.

The reading of the input via ARGF as well as the interaction with the
twitter api via the twitter gem are stubbed.

It would have been possible to stub at a lower level by using webmock to stub the
http request. However since the real http requests are abstracted away
from the app by th twitter gem, which itself is tested, this seems
unnecessary.