Fix individual tweets being over 280 characters
The TweetFormatter did check that each string in the tweets array
was no longer than 280 characters. However that was before the
enumerator for a tweet storm was added. So erroneously a tweet could be
280 characters + "1) " which would add up to 283 characters.

This commit adds a spec to reproduce the problem and fix it.
The solution is not ideal as it now enforces the same 277 character
limitation on single tweets as well.